Extracted the websocket client code into an interface that hopefully can be reused for integration tests
